games-simulation/openrct2: patch to prevent download in install phase
The upstream CMakeLists.txt is scripted in a way that during the install
phase it downloads the title sequences, which are .parkseq files that
have data what scenes to show on the title screen.  A patch has been
made to remove the lines that download, unzip and install those files
from the CMakeLists.txt and that function has been ported to the ebuild.

Package-Manager: Portage-2.3.5, Repoman-2.3.1
Signed-off-by: Jan Segre <jan@segre.in>